Move LoaderFactory.js to loaders folder
